Thunks/Callbacks: Fold boilerplate to Host.h
