+++ ThunkLibs/include/common/CopyableFunctions.h
@@ -1,4 +1,6 @@
 
+#pragma once
+
 #include <cstdint>
 #include <cstring>
 #include <cstdio>
